finalizado componente context
