+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { Home } from './pages/Home';
-import { TaskContextProvider } from './contexts/TaskContext';
 
 import './styles/theme.css';
 import './styles/global.css';
+import { TaskContextProvider } from './contexts/TaskContext/TaskContextProvider';
 
 export function App() {
   return (
