+++ src/components/CountDown/index.tsx
@@ -1,4 +1,4 @@
-import { useTaskContext } from '../../contexts/TaskContext';
+import { useTaskContext } from '../../contexts/TaskContext/useTaskContext';
 import styles from './styles.module.css';
 
 export function CountDown() {
