+++ src/components/MainForm/index.tsx
@@ -2,25 +2,10 @@ import { PlayCircleIcon } from 'lucide-react';
 import { Cycles } from '../Cycles';
 import { DefaultButton } from '../DefaultButton';
 import { DefaultInput } from '../DefaultInput';
-import { useTaskContext } from '../../contexts/TaskContext';
 
 export function MainForm() {
-  const { setState } = useTaskContext();
-
-  function handleClick() {
-    setState(prevState => {
-      return {
-        ...prevState,
-        formattedSecondsRemaining: '21:00',
-      };
-    });
-  }
-
   return (
     <form className='form' action=''>
-      <button onClick={handleClick} type='button'>
-        Clicar
-      </button>
       <div className='formRow'>
         <DefaultInput
           labelText='task'
