+++ src/contexts/TaskContext/TaskContextProvider.tsx
@@ -1,0 +1,17 @@
+import { useState } from 'react';
+import { TaskContext } from './TaskContext';
+import { initialTaskState } from './initialTaskState';
+
+type TaskContextProviderProps = {
+  readonly children: React.ReactNode;
+};
+
+export function TaskContextProvider({ children }: TaskContextProviderProps) {
+  const [state, setState] = useState(initialTaskState);
+
+  return (
+    <TaskContext.Provider value={{ state, setState }}>
+      {children}
+    </TaskContext.Provider>
+  );
+}
