+++ src/contexts/TaskContext/useTaskContext.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { TaskContext } from './TaskContext';
+
+export function useTaskContext() {
+  return useContext(TaskContext);
+}
